Add project controller and comments
