+++ src/shared/Types/types.ts
@@ -6,3 +6,15 @@ export type ErrorDetail = {
 	field: string;
 	message: string;
 };
+
+export type Requests =
+	| "GET"
+	| "POST"
+	| "PATCH"
+	| "PATCH"
+	| "PATCH_1"
+	| "DELETE"
+	| "UNAUTH"
+	| "FORBID"
+	| "ISE"
+	| "OTHER";
